safari-600.7-branch/LayoutTests/http/tests/websocket/tests/hybi/binary-type-expected.txt

CONSOLE MESSAGE: 'Blob' is not a valid value for binaryType; binaryType remains unchanged.
CONSOLE MESSAGE: 'ArrayBuffer' is not a valid value for binaryType; binaryType remains unchanged.
CONSOLE MESSAGE: '' is not a valid value for binaryType; binaryType remains unchanged.
Test WebSocket.binaryType attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ws.binaryType is "blob"
PASS ws.binaryType is "arraybuffer"
PASS ws.binaryType is "blob"
Set invalid values to binaryType. They should be ignored. No exception should be thrown.
PASS ws.binaryType is "blob"
PASS ws.binaryType is "blob"
PASS ws.binaryType is "blob"
PASS successfullyParsed is true

TEST COMPLETE
